modules/vehicles/src/lib/components/vehicle-card/vehicle-card.component.ts
import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
import { CommonModule } from '@angular/common';
import { Vehicle } from '../../models/vehicles.model';

@Component({
  selector: 'lib-vehicle-card',
  standalone: true,
  imports: [CommonModule],
  templateUrl: './vehicle-card.component.html',
  styleUrl: './vehicle-card.component.scss',
  changeDetection: ChangeDetectionStrategy.OnPush,
})
export class VehicleCardComponent {

  @Input() vehicle!: Vehicle;

  @Input() isLast!: boolean;
}
